requirements.txt
Flask
Flask_bootstrap
flask_sqlalchemy
flask_login
flast_wtf
wtforms_components
azure-cosmos